Wyvern for LOUD - asset reduction
This is a larger effort, intended to remove all the redundant assets that were packaged with Wyvern (over 500+).   The total number of loaded assets is now just over 150.   This is also a significant reduction in the mod size.

Audio assets are now housed elsewhere

Remapped a number of weapon and projectile assets to existing standard ones.

Revised a number of units animation assets, and weapon timings - and in general, just a lot of refactoring

Added a notepad session to allowing loading of all the blueprints and assets - this was instrumental in locating unused and duplicated assets.