LUA - Version 7.28
Release point

A significant fix for an issue which has likely been lurking for a LONG time.

I recently re-added several movement effects back into a few naval units.   Not an issue - however, it did reveal that movement effects, in general, were NOT being cleaned up - and would eventually stack to a point of causing a crash.

It only caught my attention when I began getting a lot of excessive 'particle cap' messages.   I originally attributed these to too many effects - on projectiles - and I never considered it was movement effects that were the true cause.

Discovered that the way 'effects' are stored, and cleaned up, had a personality issue - normally, we work directly with the data - but in this case, the older code was expecting the data to be referenced by the actual name of the variable - and not the variable itself.   Odd - and it's been this way for a VERY long time.