LUA - SetPrimarySeaAttackBase
This function, which determines the best position for the Sea Attack base, was using Amphibious pathfinding to make it's choice.

While this might seem useful, from the perspective of launching amphibious attacks against the goal - it wasn't right.

I now uses 'Water' pathing to determine the best base position - and the 'goal' is the water node closest to the actual goal - and not the actual goal itself.

Enhanced the debug dialogs to better evaluate this process