LUA - Sea Unit movement
We wanted to get more graceful naval unit movement, but the way that turn rates and acceleration are set, it's a fixed value.

By taking advantage of MotionHorzEvents, we can adjust both factors up, and down, to give a more natural, gradual turn rate, and acceleration, as naval units begin to move - and slow down.