LUA - DrawPath
I use this function to display two different path builds, one for transports, and another for the primary sea base.   What I wanted, was to match the draw color, with the type of path - so I added a color override to the function call.

This got me to review the one that I use for drawing the Attack Plan nodes as well

I also added a new AI Debug choice, to break out the transport paths, from the transport dialogs