UNITS - UEF T3 Transport
The Continental, being a 'unique' unit (the only T3 transport of the original game) has always gotten a bit of a pass in balancing, as most iconic unique units do.    It's not a large difference, but in this case, there is a small amount of room to narrow that, and still remain unique.