LUA - SetPrimaryNavalBase
This function determines which naval position will become the 'Primary' naval position.

Previously, this process used a rather simple 'Amphibious' path check to determine which naval position was closest to the AttackPlanner 'goal'.

There were some issues with that, since such a location was not often very ideal for the naval units, in as much as they might not be able to get there, or it just wasn't very good for naval operations.

Now, we'll use a 'Water' based check, and we'll actually use the pathing distance to determine which position might be a better choice.     This gives the AI a much more profound ability to 'wrap' his naval operations around land masses and get to an ideal location for naval bombardment.

Secondarily, we'll use similar logic in picking new Naval Defensive point positions, and no longer just from Naval DP markers, but unused Naval Area markers.

I have leveraged the 'draw' utility that I use for air transports, to demonstrate (via AI Debug) his pathing choices when he selects primary naval positions (appears as a short-lived blue pathing line) when the Display Attack Plans option is turned on in AI Debug.