UNITS - Teleport Nodes
The teleport structures, originally in LOUD_UNITS, have been revised and moved into the standard UNITS package.

This was done for one reason, really - to insure that the TELEPORTER category always exists, rather than resorting to using the category in some hard to find dummy unit.

The structures themselves have been resized, to permit a very open 'teleport area', around the structure, and to show a purple range ring (like engineers do) that denotes the range of the teleportation.