LUA - Teleport Nodes
The revised teleport structures - now known as Teleport Nodes, get an complete revision of their mechanics, usage costs, and the effects associated with them.

While I am still using the original Quantum Gateways, as a placeholder for this unit, they have been shrunken to about 1/4 of the original size.   This allows us to have a very clear and open 'teleport area'.   Likewise, the footprint has been increased, to more closely match the area of the teleportation effect.

Teleport nodes now get their range from the unit blueprint, as part of the Economy section, rather than having it hardcoded, so users can clearly see when units are in, or out, of the teleport area.

Teleport nodes now have a base 200E consumption, just for being there which cannot be turned off.

E Consumption is scaled to unit tier, and higher tiers also increase the time, by a small amount (only the highest increase is used - it's not cumulative - +10 seconds if a T4 unit is involved).

Consumption is now monitored, allowing the teleport to fail gracefully if power runs out during a transport - previously, the teleport remained incomplete ad infinitum until the economy event completed (which might be never).

If power reserves fall too low, the teleport button will disappear from the UI until a base minimum is in storage (1000E).

Visual effects have been rescaled to better match the true area of effect.

Reworked the tarmac operations, to permit the re-introduction of permanent tarmacs, and the proper use of a lifetime parameter, if specified.