LUA - AI Priority Functions
A major revision to the use of Priority Functions, to greatly reduce processing of platoon formers.

In many cases, a priority function check of the unit count requirement can eliminate a builder, from being reviewed, saving a lot of other checks and overhead processing.

Do this can often allow removing other static build conditions, from the builder, which can lead to reductions in the Build Condition table, and thus less overhead in processing that as well.