LUA - Tarmacs
I've recently discovered that tarmacs do NOT obey the rules for being directly removed - putting them in the trashbag - or tracking them - and trying to destroy them when the host units dies, simply does not work - at all.

Tarmacs are strictly governed by their lifetime, and will only disappear when that expires.   We default ALL tarmacs, that don't have a specified Lifetime values, to eight minutes (480), and we dispense with any trash handling, or tracking.

Cleaned up some class handling