LOUD_MISC - Static HP mod
I've reworked the Static HP mod to be just that - Static HP - and no longer impacts the relationship between shields and weapons.

The values are now dynamically handled in the blueprint, allow us to expand the range of choices.   I've included a number of options starting at below 1.0 - and up to 2.0, which is now the maximum limit.

I've chosen 2.0 as the limit primarily because it so greatly impacts the relationship of mobile units to static units.  2.0x is a drastic change to the overall game balance, and should be more than sufficient.