LOUD_MISC - Mod names
Added the LOUD cognomen to the standard available mods - to help users identify those which are maintained in-house, versus 3rd party mods.   Also added some LOG tags so we can see them being applied in the blueprint process.

This was how I determined that Static HP, and a couple of other mods, were being applied AFTER the threat assessment module.