BrewLAN - Mines
A revisit to completely finish the work on mines, all 4 factions get all 3 mines (12 units) revisited, to account for some functionality issues that LOUD was having with them.

The nuke mine has been downsized, and is now properly known as the 'Demolitions' mine, with about 9k of damage and a substantial range (13).

This move makes the T3 mine affordable and it now operates much like the others (T1 and T2) in that it is a proximity detonated mine.

All the cloaking, and more importantly, intel clearing functions, have been revised to insure that intel data is cleared, after a short period, after being detected by OMNI.