LUA - calculation of strength values
LOUD was not properly accounting for team size in it's calculations, especially when confronted with more than one enemy team (such as in a FFA).

This resulted in lower than anticipated strength values, which had all manner of follow-on effects, including the suppression of transport building and the early use of air scouts, which is, of course, a bad thing.

More importantly, this resulted in 'timid' behavior, often manifested by seeing large numbers of units just sitting about, because the strength values were low.

It was really necessary to bring the number of opponents into the calculations, both for strength, and production ratios, in addition to the already existing 'outnumbered' ratio.