LUA - table.merge versus table.cat
For a long time now, in the Distress Response process, I have been using table.merged to join lists of units from one squad, to units from another.

This is a case of not reading the process correctly - table.merged is NOT actually joining of two tables - in fact, it overwrites the keys in the first table, with keys from the second, and I would never have checked this very closely if it had not been for a 'stack overflow' error.

What I really want was table.cat - which joins two tables resulting in a new table with all the entries from both.

The side effect of this was clear - distress response would NOT respond properly - sending only a portion of the units, in some cases.

Simply put - if you have 4 units in one, and two units in the other, you'd expect to get a table with 6 units - but no - you were getting a table with 4 - the first two units in the first being overwritten by the two from the second.