LUA - AI - Factory Builders Land
As with Naval and Air, a great many changes to strength and pool size based conditions.

The end result of many of these changes is to better dovetail production, to the formation process, to avoid overbuilding (and thus idle units) and a more effective flow of units to the frontal positions.