LUA - AI - Engineer Factory Builders
This area gets some major changes, that better account for conditions where strength may indicate a great need for one type of factory over another - allowing LOUD to break the 'lockstep' that it was initially designed with.

Built around strength conditions and production ratios, it introduces new builders that will allow it to add new factories, not only out of lockstep, but sometimes in excess of the normal factory cap limits.

Also introduced a new builder that will construct (but not use as of yet) the Teleporter structure, in one of the rear base positions.