LUA - AI - Engineer tasks
Several strength based changes to engineer tasks, and a new builder (at DPs).   A new condition for certain shield building that recognizes either proximal threat - or the presence of Artillery, in order to defer/delay building of those additional shields.

Construction of T4 units is revised to stimulate building low-end T4 units sooner, and more continually, over the course of a game.