LUA - UNIT DeathWeapons
A fix for units like the ACU not firing their death weapons.  Unlike normal weapons - the secret for these units is often found directly in the script of the unit with a Fire() function (which is not a part of the Weapons class).

It was necessary to not only do the normal OnWeaponFired() function - but add the Fire() function call as well.