LUA - AI Strength/Production Ratios
We now incorporate the BuildRateModifier into the production ratio calculations, to account for the effect that the cheat level, and any outnumbered bonus.

The impact of the OutnumberedRatio has been modified down - we'll now use the root value of that to modify the air strength ratio, resulting in a somewhat lower value - the original number was just unrealistic.

In a similar fashion, we now add the impact of the A2G (Air to Ground) ratio, to the effect on the Air Strength ratio.   Essentially if the enemy air is comprised of dominantly ground attack units - we increase the AirRatio which can make LOUD much more aggressive with his air actions if that ratio gets high enough (maximum multiplier of 4 !)

We now exclude 'civilians' from all aspects of these calculations.