LUA - AI Expansion Builders - Naval
Ideally, when an enemy is within 12km - and land connected - we don't want to waste resources on going Naval.   This was at odds with an Outnumbered condition builder - that would rush the AI into the water.

By adding a threat/distance condition, and preventing the outnumbered builder from happening in the first 6 seconds (before the threat distance is established), we can suppress that unwanted behavior, without significantly delaying a valid entry into the water.