LUA - AI Dead Base monitor
The Dead base monitor was not properly declaring a base dead - as it would do so even if structures attached to the base were still present (excluding walls).    A base should not be declared 'dead' while active structures are still present (such as an airpad)

This was causing a follow-on issue with reinforcement platoons  that would arrive at a 'dead' base - and subsequently disband at that position - resulting in 'stranded' units.   Such units will now be re-directed to the closest suitable base position.

Cleaned up Dead Base Monitor AI Debug dialogs.