LUA - AI Strength calculation
Previous, the AI would accumulate the entire teams strength to compare against the enemy strength.    This often resulted in situations where one AI might be ignorant of it's own weakness in certain arms.

Now, the AI will take only it's own strength values and multiply it by the number of teammates it has, to arrive at it's relative strength.   This insures a greater relevance to it's own position, and not the team.    Doing so often resulted in it making bad strength based choices in situations where it was actually weak.