LUA - AI Air Unit Refit/Refuel operations
We've had a long standing issue with air units being 'stuck' on an airpad.   Identified one cause, the airpad, if it begins upgrading, causes any units on it to become stuck.

Added functionality to detect the completion of an upgrade/enhancement and detach all units at that point

Revised a lot of the refit dialog