LUA - AI RTB Functions
We've identified some issues with platoons which may be inbound to a base, which goes 'dead' during that operation.   Normally such a platoon would detect that the base had gone 'dead' while enroute, but, if the platoon was just about to arrive, or was involved in a 'basedistress' event, the units could be left in 'limbo' (army pool) and thus uncommandable.

Such units will now be properly re-directed to another base when that happens.

Revised a number of the AI Debug dialogs for both this, and BaseMonitor AI Debug dialogs