LUA - AI Air units in 'refit'
Air units sometimes get stuck in the refit platoon.   I can source this situation to the 'MovePlatoon' process, which is, for some reason, not exiting properly in some cases.

The AirUnitRefit process is expecting that task to end before it releases the air unit to the 'RefuelPool'.     What happens is the air units end up circling the destination airpad, indefinitely, fully fueled and repair.

We now check that condition, and force the exit, if the air unit is fueled and repaired.