LUA - AI Finding Naval DP position
When finding a new naval DP position, we check the attack plan, and try to locate the nearest Water node, to that goal.   However, sometimes, that goal is simply not realistically close - and the function that returns the node, replies with FALSE - in such a case, we'll abort the find process.