LUA - AI Builder Manager - PriorityFunction
While reviewing how Engineers handle PriorityFunctions, I noticed that any that resulted in a 0 priority (remove task) were not being actioned until the next 'sort' of the tasks.   This resulted in what should be 'dead' tasks, getting executed outside of parameters.

Revised some engineer dialogs

Added GetFocusArmy() to the preview of the pathing for PrimaryNavalBase selection