LUA - AI Base selection ranges
Some of the exclusion ranges, for base selection, have been lowered

Certain expansions (Land & Naval DP) will now test the current Attack Plan for positions first, before attempting the previous selection method.

This results in a more focused progression of the AI, towards it's selected enemy.

Addressed an issue where the AttackPlan Goal was not available (it was being calculated), causing an error