LUA - Category Restrictions
We're introducing a new mechanic that will recognize category restrictions at the pre-game level, and bypass the loading of those 'builders'.   This will reduce memory usage, and processing cycles - at present, the use of restrictions adds a good deal of processing overhead during the course of the game, as the AI is continually having to evaluate unit restrictions when trying to build with his engineers, or factories.

Added a new game function - CategoryRestricted() which resolves if a given category (rather than a UnitID) is restricted.

Added a new value to BuilderGroups, BuildersRestriction, which can hold the category that would prevent the loading of this group of builder tasks

You can see the above function in use, both with specific builders in a PriorityFunction(), or in the BuilderGroups data directly (useful when an entire set of builders is governed by the same restriction (ie. NUKES).