LUA - AI Expansion Builders
One of the DP builder checks was checking for the presence of SONAR or INTELLIGENCE - but was not removing AIR based entities - so Spy planes would sometimes cause this builder to fail on a location that was actually open, but the spy plane was passing by overhead.

Also resolved a marker radius value in the conditions that was different than what was used in the data - so the condition would pass, but the build would then fail, leading to some thrashing.