LUA - Platoon LandForceAILOUD
This behavior sees a number of AI Debug dialog refinements, and a fix to a situation where a platoon would get 'stuck' when it was unable to get transport to an otherwise unreachable goal position.

The number of transport calls has been reduced from 8, to 5, so the platoon will not sit 'idle' for nearly as long when transport is not available.