LUA - AI Behaviors
Some revisions to the loiter position process - this should make the air operations somewhat more dynamic in response to changing air strength.

The 'ProsecuteTarget' function, used by fighters and bombers at the moment, gets improved dialog and better handling of units when they exit to the 'refit' process (damage or fuel) or when they've finished their attack - and need to return to the loiterposition.

We've added a 'recalldelay' feature that will impose a delay on a unit, before returning it to the 'unassigned' squad (where it can get a new target).   This is important for bombers, who need this time to 're-arm' the bomb weapon, and gain some distance before attacking a new target.