LUA - AI Behavior - AirForceAI
For both fighter and bombers squadrons, we've revised the processing of the loiterposition, and the ProsecuteTarget() process (which moves individual units into the 'Attack' squad when they have a target).

In general, the searchradius, for targets, will be more aggressive as the AirRatio goes up - and the squadron size approaches it's maximum size.

Mission time gets lengthened when merges take place (increased size of squadron).

Fixed an oversight in the targetdistance that was not resolving the distance correctly, resulting in a poor approach to the target.

Added a 'weapon event' (WeaponFired()) to bombers, that will trigger the ProsecuteTarget() process to 'break off', after dropping it's bomb - causing the bomber to return to the loiterposition and re-select a new target.

This event is attached to any weapon on the bomber that is labelled as 'Bomb' or has the NeedToComputeBombDrop value, in the weapon blueprint.