LUA - AI base positions
The AI has recently been using 'fake' positions to establish bases using it's Attack Plan - but this, while certainly novel, has been causing some crash issues, as the location name is not unique.

Until such time as I make these positions unique in name, I'll suspend the ability to do this.   The problem is not so much this ability, but that the Attack Planner is not working hard enough to locate a legitimate position.

Added some additional debug dialog to notify me when a base position cannot be located, near to the 'goal' position (usually when seeking a base position for a naval DP).   The report will now indicate the range that was used.