BrewLAN - UEF T4 Railgun
The Excalibur gets a reduction in it's HP, regen & range, and it's E consumption moves up, all of these, about 10%.

It was never intended to be a 'frontline' weapon system, hence the change to it's HP, and the range we've used has always been a 'placeholder' until we had more data about it's usage and especially it's effectiveness, with respect to terrain and it's ability to penetrate multiple units.    It's clearly a superior weapon system, and we want to keep it that way, but we want to pull it in just a bit.