LUA - PlatoonHandle optimization
When a unit is assigned to a platoon, it's platoonhandle, and plan, and other information, is set by that process.

In numerous places in the code base, this is either repeated, or done out of sequence, which is how I noticed it - some units would not reflect the platoon change, when I was displaying platoon names (ie. - Refuel Pool)

I was simply able to remove these assignments in a number of places