LUA - AI Air Refit
While waiting to locate an airpad, there is a 1 second delay loop.    This delay was occurring even if the unit did find an airpad - now it only occurs if no airpad was found, allowing those that do find an airpad, to return to the Army pool immediately after repair.

A small thing - indeed - but still important.