LUA - AI BaseInPlayableArea
This function, often used in the Builders PriorityFunction() was being called without the aiBrain parameter (first value), which caused the function to report that the base was NOT in the playable area.

This, in turn, prevented over 30+ tasks to NOT run at the MAIN base (all of them engineering tasks).