LUA - Teleporting
Teleporting units are now subject to teleport blocking along the entire path of travel - and not just at the source or destination.

There is no 'warning' of a block, you must charge, and attempt to execute the teleport, to find out if you are blocked, or not - and incur the delays that go with a teleport action.

Teleport Nodes are not subject to blocking.