LUA - AI ReturnToBase
The RTB process is pretty complicated, so logging it, in detail, is pretty busy.   Since it's also visual, when I saw that it wasn't working properly, I thought I should look into it.

Fixed a minor data issue, normalized the logging (it's still a manual debug switch), and did some minor optimizations.

If Show Engineer Names is turned on, the name should accurately reflect what they are doing, the task itself, RTB, or idle and what their last job was.