LUA - AI Builder Tasks
I use PriorityFunctions() in a lot of builders, not just for the immediate accuracy they provide, but so that I can sidestep having and/or processing, a lot of Build Conditions.

With the kind of response I can get, I can set the priority value to something indicative, so a task might not be ready to go (10) whereas one that meets one condition may report 11.

Now, that difference means nothing to the actual execution (anything below 100 priority is simply ignored - 0 is removed) - but - if I'm using the AI Debug option that shows priority changes, I can follow the progress of the conditions, and the organization of the tasks, that the AI is evaluating - as the game progresses - so it's a helpful logging tool.

I've implemented this in a number of places (that I have recently been examining for efficiency) so I expect to do more of this at a later date.

I continue to move UnitCapCounts back into Build Conditions (and out of Priority Functions).