LUA - AIParseIntel
ParseIntel is a time-critical process that addresses the AI and the intel data.

The logging needed a review, not only to standardize it, but to better illustrate the process that's going on.

I normalize some fractional components now, primarily for better readability.

Addressed some tick counting values to firm up the process times