LUA - and not isolation
Having discovered a race condition that was directly a result of unclear 'not' conditions, I resolved to identify and clean up any others that might be lurking around.    It's unlikely that any of these might cause a similar issue, but it is quite likely that there were some logic errors happening that might have given odd results.