LUA - Mines
A submitted replay demonstrated that mines were not working, as intended - and would go into a deadloop when triggered.

Simply put, the weapon was cycling over the RackSalvoFiringState, and never triggering the OnWeaponFired() event - and would barf out an endless stream of projectile blueprint errors as well.

Fixed by installing a custom FireReady state for suicide weapons that pushes them directly to the OnWeaponFired() event (which results in the detonation and death of the mine).