LUA - M28 compatability
M28 makes calls to transport utilities, which relies on data used by LOUD's base manager - which M28 does not have.

This causes errors when M28 tries to return a transport to a base, for refitting.

Likewise, it then tries to submit the unit for refueling - but it doesn't have the refuel platoon in it's structure.