LUA - EconomyBuildConditions
The function CnaBuildOnMassAtRange() is a commonly used function for engineers, and it has a range feature to control both minimum and maximum range.

The function sorts the markerlist for maximum distance, but was not terminating once we passed that range, resulting in iterating over the entire markerlist every time.

By providing a breakpoint, we can cut the typical number of iterations by 75%.